feat(applist)!: added pagination
The UPS server application list endpoint returns a paginated result (7 items per page).
Without paginating, it was impossible to see any application after the first 7: a new `page` parameter has been added to the list command to allow seeing the following pages.

BREAKING CHANGE: now the list function takes a different set of parameters: instead of just the `filter` it now takes an object with keys `filter` and `page`. The syntax for
filter is the same as before, while the page is just a number representing the page number to be returned