feat(list)!: added management for the total header
The UPS /rest/applications endpojnt returns an header called 'total'
containing the total number of applications that match a given request
(that's very useful, since it returns the result only one page at a
time).

This implementation adds the ability to read that value and
implement the 'total' application management even for filters not
managed directly by the application server.

BREAKING CHANGE: This change will change the return value of the `find`
method